Bugfix: Handle slow starting mons daemon race condition.
When testing ceph on the cloud or on heavily loaded servers, the mon may not
have completed initialization before the code tested the server status.

This led to the mon_create method to fail, and raise an exception, and not set
the flag that it had completed successfully. This meant that when you reran the
command it would roll back and re-attempt to create the mon setup, producing the
same error.

Signed-off-by: Owen Synge <osynge@suse.com>