examples.cluster_buildup: keyring profile side effects
Due to the way keyring profiles work and the init scripts for rgw we need to
force users to only create rgw with a 'rgw.' prefix. The reason we dont hide
this from the user is due to both the systemd filea and rgw deployments may
exist without the prefix if the bootstrap keyring was not used in the key
creation for the rgw service.

For consistency the example has been changed for mds also.

Signed-off-by: Owen Synge <osynge@suse.com>