modules.ceph.init: Work around salt import bug
It seems salt adds external modules to the python path rather than as a
subdirectory or mangling the modules name. This can cause issues as stated in
bug report https://github.com/saltstack/salt/issues/35444

To work around this we import the __version__ of the library as this is only
present in the library and not in the salt module.

Signed-off-by: Owen Synge <osynge@suse.com>